organize qr library files
move nayuki's QR source files to their own directory, add QR lib license.
