+++ make.bat
@@ -1,2 +1,2 @@
 @ECHO OFF
-cl src/main.cpp src/qrcodegen.cpp src/webserver.cpp src/dataloader.cpp /Fe:out/flyle.exe /Fo:out/
+cl src/main.cpp src/nayuki-qr/qrcodegen.cpp src/webserver.cpp src/dataloader.cpp /Fe:out/flyle.exe /Fo:out/
