+++ src/nayuki-qr/qrcodegen.cpp
