+++ src/nayuki-qr/qrcodegen.hpp
